fix: move loom seed logic into app.services so Celery worker can import it
seeds/loom_references.py was not importable from Celery workers because the
seeds/ directory is not on their sys.path (app is installed as a package,
seeds/ is not). Moved the upsert logic to app/services/loom_seed.py which is
always importable. app/tasks/seeds.py now imports from app.services.loom_seed;
seeds/loom_references.py becomes a thin CLI wrapper calling the same function.
